refactor: renamed all files to underscore_case
