+++ tests/acceptance/test_softmaxAC_mc.py
@@ -1,9 +1,9 @@
 import os
 import pytest
 
-from ml_experiment.ExperimentDefinition import ExperimentDefinition
-from ml_experiment.DefinitionPart import DefinitionPart
-from ml_experiment.Scheduler import LocalRunConfig, Scheduler
+from ml_experiment.experiment_definition import ExperimentDefinition
+from ml_experiment.definition_part import DefinitionPart
+from ml_experiment.Scheduler import LocalRunConfig, RunSpec, Scheduler
 
 
 @pytest.fixture
